chore: rename folders and files
