+++ packages/forge/src/lib/breadcrumbs/breadcrumbs/_core.scss
