+++ packages/forge/src/lib/index.ts
@@ -21,7 +21,7 @@ import { defineAvatarComponent } from './avatar/index.js';
 import { defineBackdropComponent } from './backdrop/index.js';
 import { defineBadgeComponent } from './badge/index.js';
 import { defineBannerComponent } from './banner/index.js';
-import { defineBreadcrumbsComponent, defineCrumbComponent } from './breadcrumb/index.js';
+import { defineBreadcrumbsComponent, defineCrumbComponent } from './breadcrumbs/index.js';
 import { defineBottomSheetComponent } from './bottom-sheet/index.js';
 import { defineButtonAreaComponent } from './button-area/index.js';
 import { defineButtonToggleComponent, defineButtonToggleGroupComponent } from './button-toggle/index.js';
@@ -96,7 +96,7 @@ export * from './avatar/index.js';
 export * from './backdrop/index.js';
 export * from './badge/index.js';
 export * from './banner/index.js';
-export * from './breadcrumb/index.js';
+export * from './breadcrumbs/index.js';
 export * from './bottom-sheet/index.js';
 export * from './button-area/index.js';
 export * from './button-toggle/index.js';
